PiecesValidation and Sprites
Added the movement validation, now it's possible to cause a logical match, however the removal part doesn't happen yet. Also I add the sprites of the objects, otherwise it might not working properly with the PieceSO if I pull the project in the future without the art. Now I'll be working in the movement animation of the pieces, because you can't see if the piece is actually moving or was reverted because everything is happening in one frame.

Art credit:
https://ghostpixxells.itch.io/pixelfood
https://alexkovacsart.itch.io/free-pixel-art-foods