Cascade Falling
Now it's possible to do a match, and the board will move all the pieces downward, I might need to tweek the visuals in the future, because for now, the animation happens all together. Furthermore, the cascade doesn't happen automatically(even though it's on update) because I prefer to do some tests before let it run automatically to avoid unexpected bugs.

Refilling The Board -> Next Step