Board-Refill-Combo
Now the board does the refill process, it also detects further combos and refill again if needed.
A lot of null check was added, add the code is way bigger now.
Next week, my goal will be to add the score system, animations and split the BoardManager.Cs code, so it's easier to debug and expand in the future.