Bug-Fixes
Fixed a bug where the pieces weren't restoring their scale values, so they were invisible when getting them from the pool. Also some adjustments was added in the Dotween section, so it kills all animations when the piece is destroyed. Avoiding a huge amount of space for DotTweens sequences and single animations.